feat: Add `github-vcs` and `gitlab-vcs` submodules (#31)
* feat: Add github-vcs and gitlab-vcs submodules

The `github-vcs` and `gitlab-vcs` submodules build on the root module by
adding a `backend.tf` file to the specified repository and ensuring that
Terraform runs only after this file is created, so the initial plan
succeeds. Without this file, plans would fail—so it is always handled as
part of calling the module. Bundling this logic within the module simplifies
provisioning workspaces that use the VCS-driven workflow.

This commit also introduces `var.repository_files`, which allows consumers
to stub additional files in the workspace repository, such as a README,
default provider blocks, variables, and more. When specifying additional
files Terraform will wait for these to be created before queuing the
first run.

Additionally, we introduce `var.variable_set_names`, a list of strings
representing variable set names to attach to the workspace. This helps avoid
repetitive code when the variable sets are known in advance. The module
resolves these names to IDs using a data resource and passes them to the root
module. If variable sets are being created in the same configuration as this
module, `variable_set_ids` should still be used to provide the actual IDs.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>

---------

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>