fix: Correct file trigger logic and update default trigger patterns (#29)
* fix: Correct file trigger logic and update default trigger patterns

Updated default `trigger_patterns` logic to append `working_directory/*`
instead of `working_directory/**/*`. When `trigger_patterns` was
introduced it was a copy of how `trigger_patterns` worked, but this is
undesirable as we don't want to trigger plans in sub-directories;
sometimes sub-directories could be linked to other workspaces.

Fixed conditional logic for enabling `file_triggers_enabled` based on
whether patterns or prefixes are actually set.

Changed default for `trigger_prefixes` from `null` to `[]` for easier
evaluation. This should make no difference so not considering it a
breaking change as the result is the same when not set.

Added tests to help with testing change in logic.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>

---------

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>